refactor: extract pgrest row sanitization to shared module
